feat(Editor) Migrated the old css-style-loader to the webpack styled loaders
